Corrected unlimited memory setting for batching to use proper type

+++ libtensor/block_sparse/sparse_btensor.h
@@ -258,7 +258,7 @@ void sparse_btensor<N,T>::assign(const expr::expr_rhs<N, T>& rhs, const expr::la
     bp.get_direct_bispace_grps(direct_bispace_grps);
     bp.get_batched_subspace_grps(batched_subspace_grps);
     batch_list_builder blb(direct_bispace_grps,batched_subspace_grps);
-    size_t mem_avail = (m_mr != NULL) ? m_mr->get_mem_avail() : std::numeric_limits<double>::max();
+    size_t mem_avail = (m_mr != NULL) ? m_mr->get_mem_avail() : std::numeric_limits<size_t>::max();
     idx_pair_list batch_list = blb.get_batch_list(mem_avail/sizeof(T));
     bp.set_batch_info(batched_subspace_grps,batch_list);
